New destroy action

+++ web/opennaas-web/src/main/java/org/opennaas/web/actions/DestroyAction.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import org.apache.log4j.Logger;
 import org.apache.struts2.interceptor.SessionAware;
 import org.opennaas.web.ws.OpennaasClient;
 import org.opennaas.ws.ActionException_Exception;
@@ -28,8 +29,14 @@ import com.opensymphony.xwork2.ActionSupport;
  */
 public class DestroyAction extends ActionSupport implements SessionAware {
 
-	private Map<String, Object>		session;
-	private IResourceManagerService	resourceManagerService;
+	private static final Logger				LOGGER	= Logger.getLogger(DestroyAction.class);
+	private Map<String, Object>				session;
+	private IResourceManagerService			resourceManagerService;
+	private INetQueueCapabilityService		queueService;
+	private IChassisCapabilityService		chassisCapab;
+	private IQueueManagerCapabilityService	queueCapab;
+	private IGRETunnelCapabilityService		greCapab;
+	private IL2BoDCapabilityService			l2BoDCapabilityService;
 
 	@Override
 	public void setSession(Map<String, Object> session) {
@@ -65,44 +72,95 @@ public class DestroyAction extends ActionSupport implements SessionAware {
 	}
 
 	/**
+	 * @throws CapabilityException_Exception
+	 */
+	private void deactivateOSPF() throws CapabilityException_Exception {
+		queueService = OpennaasClient.getNetQueueCapabilityService();
+		try {
+			if (session.get(getText("network.name")) != null) {
+				String networkId = ((ResourceIdentifier) session.get(getText("network.name"))).getId();
+				INetOSPFCapabilityService capabilityService = OpennaasClient.getNetOSPFCapabilityService();
+				capabilityService.deactivateOSPF(networkId);
+
+				queueService.execute(networkId);
+			}
+		} catch (Exception e) {
+			LOGGER.error("Can't deactivate ospf.");
+		}
+
+	}
+
+	/**
 	 * @throws ResourceException_Exception
 	 * @throws CapabilityException_Exception
 	 * @throws ActionException_Exception
 	 * @throws ProtocolException_Exception
 	 */
 	private void removeLR() throws ResourceException_Exception, CapabilityException_Exception, ActionException_Exception, ProtocolException_Exception {
-		resourceManagerService = OpennaasClient.getResourceManagerService();
+		try {
+			chassisCapab = OpennaasClient.getChassisCapabilityService();
+			resourceManagerService = OpennaasClient.getResourceManagerService();
+			queueCapab = OpennaasClient.getQueueManagerCapabilityService();
 
-		ResourceIdentifier lrUnic = (ResourceIdentifier) session.get(getText("unic.lrouter.name"));
-		ResourceIdentifier lrGSN = (ResourceIdentifier) session.get(getText("myre.lrouter.name"));
-		ResourceIdentifier lrMyre = (ResourceIdentifier) session.get(getText("gsn.lrouter.name"));
+			ResourceIdentifier lrUnic = (ResourceIdentifier) session.get(getText("unic.lrouter.name"));
+			ResourceIdentifier lrGSN = (ResourceIdentifier) session.get(getText("myre.lrouter.name"));
+			ResourceIdentifier lrMyre = (ResourceIdentifier) session.get(getText("gsn.lrouter.name"));
 
-		resourceManagerService.stopResource(lrUnic);
-		resourceManagerService.stopResource(lrMyre);
-		resourceManagerService.stopResource(lrGSN);
+			try {
+				if (lrUnic != null) {
+					resourceManagerService.stopResource(lrUnic);
+					resourceManagerService.removeResourceById(lrUnic.getId());
 
-		resourceManagerService.removeResourceById(lrUnic.getId());
-		resourceManagerService.removeResourceById(lrMyre.getId());
-		resourceManagerService.removeResourceById(lrGSN.getId());
+					if (session.get(getText("unic.router.name")) != null) {
+						ComputerSystem router = new ComputerSystem();
+						String routerUnicId = ((ResourceIdentifier) session.get(getText("unic.router.name"))).getId();
+						router.setName(getText("unic.lrouter.name"));
+						chassisCapab.deleteLogicalRouter(routerUnicId, router);
+						queueCapab.execute(routerUnicId);
+					}
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't delete " + getText("unic.lrouter.name") + " LR");
+			}
 
-		String routerUnicId = ((ResourceIdentifier) session.get(getText("unic.router.name"))).getId();
-		String routerMyreId = ((ResourceIdentifier) session.get(getText("myre.router.name"))).getId();
-		String routerGSNId = ((ResourceIdentifier) session.get(getText("gsn.router.name"))).getId();
+			try {
+				if (lrMyre != null) {
+					resourceManagerService.stopResource(lrMyre);
+					resourceManagerService.removeResourceById(lrMyre.getId());
 
-		IChassisCapabilityService chassisCapab = OpennaasClient.getChassisCapabilityService();
+					if (session.get(getText("myre.router.name")) != null) {
+						ComputerSystem router = new ComputerSystem();
+						String routerMyreId = ((ResourceIdentifier) session.get(getText("myre.router.name"))).getId();
+						router.setName(getText("myre.lrouter.name"));
+						chassisCapab.deleteLogicalRouter(routerMyreId, router);
+						queueCapab.execute(routerMyreId);
+					}
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't delete " + getText("myre.lrouter.name") + " LR");
+			}
 
-		ComputerSystem router = new ComputerSystem();
-		router.setName(getText("unic.lrouter.name"));
-		chassisCapab.deleteLogicalRouter(routerUnicId, router);
-		router.setName(getText("myre.lrouter.name"));
-		chassisCapab.deleteLogicalRouter(routerMyreId, router);
-		router.setName(getText("gsn.lrouter.name"));
-		chassisCapab.deleteLogicalRouter(routerGSNId, router);
+			try {
+				if (lrMyre != null) {
+					resourceManagerService.stopResource(lrGSN);
+					resourceManagerService.removeResourceById(lrGSN.getId());
 
-		IQueueManagerCapabilityService queueCapab = OpennaasClient.getQueueManagerCapabilityService();
-		queueCapab.execute(routerUnicId);
-		queueCapab.execute(routerMyreId);
-		queueCapab.execute(routerGSNId);
+					if (session.get(getText("gsn.router.name")) != null) {
+						ComputerSystem router = new ComputerSystem();
+						String routerGSNId = ((ResourceIdentifier) session.get(getText("gsn.router.name"))).getId();
+						router.setName(getText("gsn.lrouter.name"));
+						chassisCapab.deleteLogicalRouter(routerGSNId, router);
+						queueCapab.execute(routerGSNId);
+					}
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't delete " + getText("gsn.router.name") + " LR");
+			}
+
+		} catch (Exception e) {
+			LOGGER.error("Can't delete all logical routers.");
+		}
+
 	}
 
 	/**
@@ -111,29 +169,22 @@ public class DestroyAction extends ActionSupport implements SessionAware {
 	 * @throws ProtocolException_Exception
 	 */
 	private void removeGRE() throws CapabilityException_Exception, ActionException_Exception, ProtocolException_Exception {
-		String lrMyreId = ((ResourceIdentifier) session.get(getText("myre.lrouter.name"))).getId();
+		try {
+			queueCapab = OpennaasClient.getQueueManagerCapabilityService();
+			greCapab = OpennaasClient.getGRETunnelCapabilityService();
+			if (session.get(getText("myre.lrouter.name")) != null) {
+				GreTunnelService greTunnelService = new GreTunnelService();
+				String lrMyreId = ((ResourceIdentifier) session.get(getText("myre.lrouter.name"))).getId();
 
-		GreTunnelService greTunnelService = new GreTunnelService();
+				greTunnelService.setName(getText("myre.iface.gre"));
+				greCapab.deleteGRETunnel(lrMyreId, greTunnelService);
 
-		IGRETunnelCapabilityService greCapab = OpennaasClient.getGRETunnelCapabilityService();
+				queueCapab.execute(lrMyreId);
+			}
+		} catch (Exception e) {
+			LOGGER.error("Can't delete GRE.");
+		}
 
-		greTunnelService.setName(getText("myre.iface.gre"));
-		greCapab.deleteGRETunnel(lrMyreId, greTunnelService);
-
-		IQueueManagerCapabilityService queueCapab = OpennaasClient.getQueueManagerCapabilityService();
-		queueCapab.execute(lrMyreId);
-	}
-
-	/**
-	 * @throws CapabilityException_Exception
-	 */
-	private void deactivateOSPF() throws CapabilityException_Exception {
-		String networkId = ((ResourceIdentifier) session.get(getText("network.name"))).getId();
-		INetOSPFCapabilityService capabilityService = OpennaasClient.getNetOSPFCapabilityService();
-		capabilityService.deactivateOSPF(networkId);
-
-		INetQueueCapabilityService queueService = OpennaasClient.getNetQueueCapabilityService();
-		queueService.execute(networkId);
 	}
 
 	/**
@@ -142,23 +193,47 @@ public class DestroyAction extends ActionSupport implements SessionAware {
 	 * @throws CapabilityException_Exception
 	 */
 	private void shutDownAutobahn() throws CapabilityException_Exception, ResourceException_Exception {
-		IL2BoDCapabilityService l2BoDCapabilityService = OpennaasClient.getL2BoDCapabilityService();
-		String autbahnId = ((ResourceIdentifier) session.get(getText("autobahn.bod.name"))).getId();
+		try {
+			l2BoDCapabilityService = OpennaasClient.getL2BoDCapabilityService();
+			if (session.get(getText("autobahn.bod.name")) != null) {
+				String autbahnId = ((ResourceIdentifier) session.get(getText("autobahn.bod.name"))).getId();
 
-		List<String> list1 = new ArrayList<String>();
-		list1.add(getText("autobahn.connection1.interface1"));
-		list1.add(getText("autobahn.connection1.interface2"));
-		l2BoDCapabilityService.shutDownConnection(autbahnId, list1);
+				try {
+					List<String> list1 = new ArrayList<String>();
+					list1.add(getText("autobahn.connection1.interface1"));
+					list1.add(getText("autobahn.connection1.interface2"));
+					l2BoDCapabilityService.shutDownConnection(autbahnId, list1);
+				} catch (Exception e) {
+					LOGGER.error("Can't shut down autobahn connection 1 - "
+							+ getText("autobahn.connection1.interface1") + " - "
+							+ getText("autobahn.connection1.interface2"));
+				}
 
-		List<String> list2 = new ArrayList<String>();
-		list2.add(getText("autobahn.connection2.interface1"));
-		list2.add(getText("autobahn.connection2.interface2"));
-		l2BoDCapabilityService.shutDownConnection(autbahnId, list2);
+				try {
+					List<String> list2 = new ArrayList<String>();
+					list2.add(getText("autobahn.connection2.interface1"));
+					list2.add(getText("autobahn.connection2.interface2"));
+					l2BoDCapabilityService.shutDownConnection(autbahnId, list2);
+				} catch (Exception e) {
+					LOGGER.error("Can't shut down autobahn connection 2 - "
+							+ getText("autobahn.connection2.interface1") + " - "
+							+ getText("autobahn.connection2.interface2"));
+				}
 
-		List<String> list3 = new ArrayList<String>();
-		list3.add(getText("autobahn.connection3.interface1"));
-		list3.add(getText("autobahn.connection3.interface2"));
-		l2BoDCapabilityService.shutDownConnection(autbahnId, list3);
+				try {
+					List<String> list3 = new ArrayList<String>();
+					list3.add(getText("autobahn.connection3.interface1"));
+					list3.add(getText("autobahn.connection3.interface2"));
+					l2BoDCapabilityService.shutDownConnection(autbahnId, list3);
+				} catch (Exception e) {
+					LOGGER.error("Can't shut down autobahn connection 3 - "
+							+ getText("autobahn.connection3.interface1") + " - "
+							+ getText("autobahn.connection3.interface2"));
+				}
+			}
+		} catch (Exception e) {
+			LOGGER.error("Can't shut down autobahn.");
+		}
 	}
 
 	/**
@@ -167,22 +242,61 @@ public class DestroyAction extends ActionSupport implements SessionAware {
 	 * @throws ProtocolException_Exception
 	 */
 	private void removeResources() throws CapabilityException_Exception, ResourceException_Exception, ProtocolException_Exception {
-		resourceManagerService = OpennaasClient.getResourceManagerService();
+		try {
+			resourceManagerService = OpennaasClient.getResourceManagerService();
 
-		ResourceIdentifier routerUnic = (ResourceIdentifier) session.get(getText("unic.router.name"));
-		ResourceIdentifier routerGSN = (ResourceIdentifier) session.get(getText("gsn.router.name"));
-		ResourceIdentifier routerMyre = (ResourceIdentifier) session.get(getText("myre.router.name"));
-		ResourceIdentifier networkDemo = (ResourceIdentifier) session.get(getText("network.name"));
+			try {
+				if (session.get(getText("unic.router.name")) != null) {
+					ResourceIdentifier routerUnic = (ResourceIdentifier) session.get(getText("unic.router.name"));
+					resourceManagerService.stopResource(routerUnic);
+					resourceManagerService.removeResourceById(routerUnic.getId());
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't remove or stop " + getText("unic.router.name"));
+			}
 
-		resourceManagerService.stopResource(routerUnic);
-		resourceManagerService.stopResource(routerGSN);
-		resourceManagerService.stopResource(routerMyre);
-		resourceManagerService.stopResource(networkDemo);
+			try {
+				if (session.get(getText("gsn.router.name")) != null) {
+					ResourceIdentifier routerGSN = (ResourceIdentifier) session.get(getText("gsn.router.name"));
+					resourceManagerService.stopResource(routerGSN);
+					resourceManagerService.removeResourceById(routerGSN.getId());
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't remove or stop " + getText("gsn.router.name"));
+			}
 
-		resourceManagerService.removeResourceById(routerUnic.getId());
-		resourceManagerService.removeResourceById(routerGSN.getId());
-		resourceManagerService.removeResourceById(routerMyre.getId());
-		resourceManagerService.removeResourceById(networkDemo.getId());
+			try {
+				if (session.get(getText("myre.router.name")) != null) {
+					ResourceIdentifier routerMyre = (ResourceIdentifier) session.get(getText("myre.router.name"));
+					resourceManagerService.stopResource(routerMyre);
+					resourceManagerService.removeResourceById(routerMyre.getId());
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't remove or stop " + getText("myre.router.name"));
+			}
+
+			try {
+				if (session.get(getText("network.name")) != null) {
+					ResourceIdentifier networkDemo = (ResourceIdentifier) session.get(getText("network.name"));
+					resourceManagerService.stopResource(networkDemo);
+					resourceManagerService.removeResourceById(networkDemo.getId());
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't remove or stop " + getText("network.name"));
+			}
+
+			try {
+				if (session.get(getText("autobahn.bod.name")) != null) {
+					ResourceIdentifier autobahnDemo = (ResourceIdentifier) session.get(getText("autobahn.bod.name"));
+					resourceManagerService.stopResource(autobahnDemo);
+					resourceManagerService.removeResourceById(autobahnDemo.getId());
+				}
+			} catch (Exception e) {
+				LOGGER.error("Can't remove or stop " + getText("network.name"));
+			}
+
+		} catch (Exception e) {
+			LOGGER.error("Can't remove resources.");
+		}
 	}
-
 }
